Moved sleep.py to main directory as functions folder does not have any use
Might revert back later
